+++ UWP/Helpers/GraphHelper.cs
@@ -4,14 +4,14 @@ using System.Collections.Generic;
 
 namespace UWP.Helpers {
     class GraphHelper {
-        internal static (float Min, float Max) GetMinMaxOfArray(List<float> historic) {
+        internal static (double Min, double Max) GetMinMaxOfArray(List<double> historic) {
             if (historic.Count == 0)
                 return (0, 10);
 			
-            float min = historic[0];
-            float max = historic[0];
+            double min = historic[0];
+            double max = historic[0];
 
-            foreach (float h in historic) {
+            foreach (double h in historic) {
                 if (h < min)
                     min = h;
                 else if (h > max)
@@ -20,8 +20,8 @@ namespace UWP.Helpers {
 
             double diff = max - min;
 
-            min -= (float)(diff * 0.2);
-            max += (float)(diff * 0.1);
+            min -= (double)(diff * 0.2);
+            max += (double)(diff * 0.1);
 
             if (min < 0)
                 min = 0;
